HTTPS/SSL Proxy support!
This enables the use of a proxy for https requests.  To do so pass the new --httpproxy-ssl (or %C) option.
Note: This sends requests to the proxy over HTTP not HTTPS, most browsers use HTTPS and the connect command.
All proxies may not work with this method, Fiddler does.
Still needs full documentation and UI options added.